# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "User API",
		"description": "Tandain User API",
		"version": "0.1.1",
		"license": {
			"name": "MIT,",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3001",
			"description": "Local server for development"
		}
	],
	"components": {
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},

		"schemas": {
			"BASE_ERROR_RESPONSE": {
				"properties": {
					"name": {
						"type": "string"
					},
					"code": {
						"type": "integer",
						"format": "int64"
					},
					"message": {
						"type": "string"
					}
				}
			}
		},

		"examples": {
			"INTERNAL_SERVER_ERROR": {
				"description": "Something went wrong with server.",
				"value": {
					"name": "INTERNAL_SERVER_ERROR",
					"code": 500,
					"message": "Internal Server Error"
				}
			},
			"INVALID_GRANT": {
				"description": "Authentication code is incorrect, invalid, or expired.",
				"value": {
					"name": "Invalid Grant",
					"code": 400,
					"message": "Required parameter 'code' is expired or invalid"
				}
			},
			"INVALID_REFRESH_TOKEN": {
				"description": "Refresh token is incorrect or invalid.",
				"value": {
					"name": "INVALID_REFRESH_TOKEN",
					"code": 400,
					"message": "Required parameter 'refresh_token' is invalid"
				}
			},
			"REFRESH_TOKEN_EXPIRED": {
				"description": "Refresh token is no longer valid.",
				"value": {
					"name": "REFRESH_TOKEN_EXPIRED",
					"code": 400,
					"message": "Required parameter 'refresh_token' is expired"
				}
			},
			"USER_NOT_FOUND": {
				"description": "The user corresponding to the parameter was not found.",
				"value": {
					"name": "USER_NOT_FOUND",
					"code": 400,
					"message": "User is not found"
				}
			}
		}
	},
	"paths": {
		"/api/v1/auth/login": {
			"post": {
				"summary": "Return user credentials",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"code": {
										"type": "string",
										"example": "4/0AX4XfWjvrNROKEUeD6KJipNcKqV2wswwZkEGxCm7ChHs2ZCrSpl2h9-FIx7OSaRMo0w3Pg"
									},
									"redirectUri": {
										"type": "string",
										"example": "http://localhost:3000/auth/google-oauth"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Logged in successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Logged in successfully"
										},
										"id_token": {
											"type": "string",
											"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiIzMDAxIiwic3ViIjoxLCJleHAiOjE2NTk4NzY3MTIyMDMsImF1ZCI6IjMwMDEiLCJuYW1lIjoiUGVpdHJhIEVyZGkiLCJlbWFpbCI6InBlaXRyYWVAZ21haWwuY29tIiwiaWF0IjoxNjU5ODczMTEyfQ.SVu5jVWTPxy-eBTQWTq58iba7uuryeNNRgZYAWwfqyA"
										},
										"expiry_date": {
											"type": "integer",
											"format": "int64",
											"example": 1660807905766
										},
										"refresh_token": {
											"type": "string",
											"example": "avvRXUia+DXMtm77YxnueZh3clI7DrOjwDUNdqjb78JDfaxSar7yq2c4BB/sql5r"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Authentication code is expired or invalid",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BASE_ERROR_RESPONSE"
								},
								"examples": {
									"INVALID_GRANT": {
										"$ref": "#/components/examples/INVALID_GRANT"
									}
								}
							}
						}
					},
					"500": {
						"description": "Something went wrong with server",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BASE_ERROR_RESPONSE"
								},
								"examples": {
									"INTERNAL_SERVER_ERROR": {
										"$ref": "#/components/examples/INTERNAL_SERVER_ERROR"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/refresh": {
			"post": {
				"summary": "Exchange a refresh token for an ID token",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"refresh_token": {
										"type": "string",
										"example": "avvRXUia+DXMtm77YxnueZh3clI7DrOjwDUNdqjb78JDfaxSar7yq2c4BB/sql5r"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "A successful request",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id_token": {
											"type": "string",
											"example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiIzMDAxIiwic3ViIjoxLCJleHAiOjE2NTk4NzY3MTIyMDMsImF1ZCI6IjMwMDEiLCJuYW1lIjoiUGVpdHJhIEVyZGkiLCJlbWFpbCI6InBlaXRyYWVAZ21haWwuY29tIiwiaWF0IjoxNjU5ODczMTEyfQ.SVu5jVWTPxy-eBTQWTq58iba7uuryeNNRgZYAWwfqyA"
										},
										"expiry_date": {
											"type": "integer",
											"format": "int64",
											"example": 1659876724659
										},
										"refresh_token": {
											"type": "string",
											"example": "avvRXUia+DXMtm77YxnueZh3clI7DrOjwDUNdqjb78JDfaxSar7yq2c4BB/sql5r"
										},
										"message": {
											"type": "string",
											"example": "Refresh token successfully"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "An invalid refresh token is provided",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BASE_ERROR_RESPONSE"
								},
								"examples": {
									"INVALID_REFRESH_TOKEN": {
										"$ref": "#/components/examples/INVALID_REFRESH_TOKEN"
									},
									"REFRESH_TOKEN_EXPIRED": {
										"$ref": "#/components/examples/REFRESH_TOKEN_EXPIRED"
									},
									"USER_NOT_FOUND": {
										"$ref": "#/components/examples/USER_NOT_FOUND"
									}
								}
							}
						}
					},
					"500": {
						"description": "Something went wrong with server",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BASE_ERROR_RESPONSE"
								},
								"examples": {
									"INTERNAL_SERVER_ERROR": {
										"$ref": "#/components/examples/INTERNAL_SERVER_ERROR"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/logout": {
			"post": {
				"summary": "Revoke user authentication",
				"security": [
					{
						"BearerAuth": ["write"]
					}
				],
				"responses": {
					"302": {
						"description": "Log out user and redirect them to home page"
					},
					"500": {
						"description": "Something went wrong with server",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BASE_ERROR_RESPONSE"
								},
								"examples": {
									"INTERNAL_SERVER_ERROR": {
										"$ref": "#/components/examples/INTERNAL_SERVER_ERROR"
									}
								}
							}
						}
					}
				}
			}
		}
	}
}